Move Elasticsearch code under backend package
