+++ src/test/java/io/jenkins/plugins/opentelemetry/backend/elastic/ElasticsearchContainer.java
@@ -2,7 +2,7 @@
  * Copyright The Original Author or Authors
  * SPDX-License-Identifier: Apache-2.0
  */
-package io.jenkins.plugins.opentelemetry.job.log;
+package io.jenkins.plugins.opentelemetry.backend.elastic;
 
 import io.jenkins.plugins.opentelemetry.semconv.OpenTelemetryTracesSemanticConventions;
 import org.apache.http.HttpHost;
